+++ src/test/java/com/contentful/java/cda/SyncTest.java
@@ -16,7 +16,7 @@ import static com.google.common.truth.Truth.assertThat;
 public class SyncTest extends BaseTest {
   @Test @Enqueue({
       "demo/sync_initial_p1.json", "demo/sync_initial_p2.json",
-      "demo/space.json", "demo/content_types.json",
+      "demo/locales.json", "demo/content_types.json",
       "demo/sync_update_p1.json", "demo/sync_update_p2.json"
   })
   public void sync() throws Exception {
@@ -95,8 +95,8 @@ public class SyncTest extends BaseTest {
   }
 
   @SuppressWarnings("unchecked") @Test @Enqueue(defaults = {}, value = {
-      "shallow/space.json", "shallow/types.json", "shallow/initial.json",
-      "shallow/space.json", "shallow/types.json", "shallow/update.json"
+      "shallow/locales.json", "shallow/types.json", "shallow/initial.json",
+      "shallow/locales.json", "shallow/types.json", "shallow/update.json"
   })
   public void testRawFields() throws Exception {
     SynchronizedSpace space = client.sync().fetch();
@@ -139,7 +139,7 @@ public class SyncTest extends BaseTest {
   @Enqueue(
       defaults = {},
       value = {
-          "links_invalid/space.json",
+          "links_invalid/locales.json",
           "links_invalid/content_types.json",
           "links_invalid/sync_initial.json"
       }
@@ -150,7 +150,7 @@ public class SyncTest extends BaseTest {
 
   @Test @Enqueue({
       "demo/sync_initial_preview_p1.json", "demo/sync_initial_preview_p2.json",
-      "demo/space.json", "demo/content_types.json"
+      "demo/locales.json", "demo/content_types.json"
   })
   public void syncingInPreviewWithTokenSyncsInitial() throws Exception {
     client = createPreviewClient();
@@ -162,9 +162,9 @@ public class SyncTest extends BaseTest {
   @Test
   @Enqueue({
       "demo/sync_initial_preview_p1.json", "demo/sync_initial_preview_p2.json",
-      "demo/space.json", "demo/content_types.json",
+      "demo/locales.json", "demo/content_types.json",
       "demo/sync_initial_preview_p1.json", "demo/sync_initial_preview_p2.json",
-      "demo/space.json", "demo/content_types.json"
+      "demo/locales.json", "demo/content_types.json"
   })
   public void syncingInPreviewWithPreviousSpaceSyncsInitial() throws Exception {
     client = createPreviewClient();
@@ -180,7 +180,7 @@ public class SyncTest extends BaseTest {
   @Enqueue({
       "demo/sync_initial_preview_p1.json",
       "demo/sync_initial_preview_p2.json",
-      "demo/space.json"})
+      "demo/locales.json"})
   public void syncingWithPreviewWorks() throws Exception {
     client = createPreviewClient();
     final SynchronizedSpace space = client.sync().fetch();